+++ l10n_hr_account_invoice_tips_fiskal/models/fiskal_mixin.py
@@ -1,12 +1,34 @@
-from odoo import _, models
+from datetime import datetime
+from odoo import _, fields, models
 from odoo.addons.l10n_hr_account_fiskal.fiskal import fiskal
+from odoo.fields import Datetime
 from odoo.tools import float_compare
 from odoo.exceptions import ValidationError
+from odoo.addons.l10n_hr_base.models.res_company import RACUN_DATETIME_FORMAT
 
 
 class FiscalFiscalMixin(models.AbstractModel):
     """"Extend to support tips fiscalization"""
     _inherit = "l10n.hr.fiskal.mixin"
+
+    l10n_hr_fiskal_tip_user_id = fields.Many2one(
+        comodel_name="res.partner",
+        string="Tip Fiskal User",
+        domain=lambda self: self._get_l10n_hr_fiskal_user_id_domain(),
+        ondelete='restrict',
+        copy=False,
+        help="User who fiskalized invoice tips (for internal use)",
+    )
+    l10n_hr_fiskal_tip_refund_user_id = fields.Many2one(
+        comodel_name="res.partner",
+        string="Tip Refund Fiskal User",
+        domain=lambda self: self._get_l10n_hr_fiskal_user_id_domain(),
+        ondelete='restrict',
+        copy=False,
+        help="User who refunded fiskalized invoice tips (for internal use)",
+    )
+    l10n_hr_fiskal_tip_date = fields.Datetime(string="Tip Fiskalization Date", copy=False)
+    l10n_hr_fiskal_tip_refund_date = fields.Datetime(string="Tip Refund Fiskalization Date", copy=False)
 
     def _l10n_hr_fiscalization_needed(self, message_type):
         """"Check if tips should be fiskalized"""
@@ -61,15 +83,34 @@ class FiscalFiscalMixin(models.AbstractModel):
             response.PorukaOdgovora.SifraPoruke == 'p002'
         ):
             self.l10n_hr_napojnica_is_fiscalized = True
+            if not self.l10n_hr_fiskal_tip_date:
+                self.l10n_hr_fiskal_tip_date = Datetime.now()
+            # write user who fiskalized tip if user is not set
+            if not self.l10n_hr_fiskal_tip_user_id:
+                self.l10n_hr_fiskal_tip_user_id = self.l10n_hr_fiskal_user_id.id
+            # update refund fields
             if self._context.get('l10n_hr_tip_refund'):
                 self.l10n_hr_napojnica_refund_fiscalized = True
-
+                if not self.l10n_hr_fiskal_tip_refund_date:
+                    self.l10n_hr_fiskal_tip_refund_date = Datetime.now()
+                # update user who refunded tip if user is not set
+                if not self.l10n_hr_fiskal_tip_refund_user_id:
+                    self.l10n_hr_fiskal_tip_refund_user_id = self.l10n_hr_fiskal_tip_user_id.id
         return res
 
     def _check_tips_before_fiskalize(self, msg_type):
         """Check if ZKI, napojnica_iznos, and napojnica_payment_type are set before fisk."""
         if msg_type != 'napojnica':
             return
+
+        # NOTE: check if more than 2 days passed after invoice was fiskalized
+        current_time = fields.Datetime.context_timestamp(self, datetime.now()).replace(tzinfo=None)
+        vrijeme_izdavanja = datetime.strptime(
+            self.l10n_hr_vrijeme_izdavanja, RACUN_DATETIME_FORMAT
+        )
+        if (current_time-vrijeme_izdavanja).days >= 2:
+            raise ValidationError(_("It is not possible to fiskalize the tip because more than 2 days "
+                "have passed since the invoice was issued."))
         if self.l10n_hr_napojnica_refund_fiscalized:
             raise ValidationError(_("Tip refund is already fiskalized."))
         # NOTE: skip check if refund is called