+++ l10n_hr_account_invoice_tips_fiskal/readme/HISTORY.rst
@@ -1,3 +1,10 @@
+17.0.1.1.0 (2024-09-30)
+~~~~~~~~~~~~~~~~~~~~~~~
+
+* Added info field in Tips tab who and when fiskalized tip or tip refund
+* Added check to raise error if user tries to fiskalize tip 2 days after invoice was issued
+
+
 17.0.1.0.0 (2024-09-30)
 ~~~~~~~~~~~~~~~~~~~~~~~